vfio: move remaining parts from core to container
Move the remaining functions from vfio/core into vfio/container.
