+++ src/vfio/container.c
@@ -45,6 +45,8 @@
 
 #include "container.h"
 
+struct vfio_container vfio_default_container = {};
+
 static void UNNEEDED __unmap_mapping(void *opaque, struct iova_mapping *m)
 {
 	struct vfio_container *vfio = opaque;
@@ -346,6 +348,24 @@ int vfio_init_container(struct vfio_container *vfio)
 		return -1;
 	}
 	return 0;
+}
+
+struct vfio_container *vfio_new(void)
+{
+	struct vfio_container *vfio = znew_t(struct vfio_container, 1);
+
+	if (vfio_init_container(vfio) < 0) {
+		free(vfio);
+		return NULL;
+	}
+
+	return vfio;
+}
+
+static void __attribute__((constructor)) open_default_container(void)
+{
+	if (vfio_init_container(&vfio_default_container))
+		log_debug("default container not initialized\n");
 }
 
 int vfio_do_map_dma(struct vfio_container *vfio, void *vaddr, size_t len, uint64_t iova)
