+++ src/vfio/container.h
@@ -36,6 +36,5 @@ struct vfio_container {
 extern struct vfio_container vfio_default_container;
 
 int vfio_get_device_fd(struct vfio_container *vfio, const char *bdf);
-int vfio_init_container(struct vfio_container *vfio);
 int vfio_do_map_dma(struct vfio_container *vfio, void *vaddr, size_t len, uint64_t iova);
 int vfio_do_unmap_dma(struct vfio_container *vfio, size_t len, uint64_t iova);
