--- src/vfio/core.c
@@ -1,50 +1,0 @@
-// SPDX-License-Identifier: LGPL-2.1-or-later or MIT
-
-/*
- * This file is part of libvfn.
- *
- * Copyright (C) 2022 The libvfn Authors. All Rights Reserved.
- *
- * This library (libvfn) is dual licensed under the GNU Lesser General
- * Public License version 2.1 or later or the MIT license. See the
- * COPYING and LICENSE files for more information.
- */
-
-#include <errno.h>
-#include <inttypes.h>
-#include <stdarg.h>
-#include <stdbool.h>
-#include <stdio.h>
-#include <stdlib.h>
-#include <pthread.h>
-
-#include <sys/mman.h>
-
-#include "vfn/support/atomic.h"
-#include "vfn/support/compiler.h"
-#include "vfn/support/log.h"
-#include "vfn/support/mem.h"
-
-#include "vfn/vfio/container.h"
-
-#include "container.h"
-
-struct vfio_container vfio_default_container = {};
-
-struct vfio_container *vfio_new(void)
-{
-	struct vfio_container *vfio = znew_t(struct vfio_container, 1);
-
-	if (vfio_init_container(vfio) < 0) {
-		free(vfio);
-		return NULL;
-	}
-
-	return vfio;
-}
-
-static void __attribute__((constructor)) open_default_container(void)
-{
-	if (vfio_init_container(&vfio_default_container))
-		log_debug("default container not initialized\n");
-}
